Fresh implementation of propertree
A mostly backwards compatible re-implementation of
propertree. The original implementation remains for
now and is marked as deprecated. It will be removed
at some future date. A key new feature introduced
here is native support for logical grouping such
as 'and', 'or' etc as well as literal types str, int
etc.